\name{funcionauxiliar}
\alias{funcionauxiliar}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
fauxiliar}
\description{
Intermediate auxiliar function used for allocation rule in MCT (two-lines).
}
\usage{
funcionauxiliar(permute, costes)
}
\arguments{
  \item{permute}{
Permutation.
}
  \item{costes}{
Vector. Costs to each agent.
}
}